+++ doc/changelog.rst
@@ -3,6 +3,17 @@ Changelog
 
 This changelog only includes the most important changes in recent updates. For a full log of all changes, please refer to git.
 
+Version 2.18.9
+--------------
+* Added the reb_serialize_particle_data function for fast access to particle data via numpy array. The full syntax is explain in the documentation. Here is a short example: 
+.. code:: python
+   
+   import numpy as np
+   a = np.zeros((sim.N,3),dtype="float64")
+   sim.serialize_particle_data(xyz=a)
+   print(a)
+
+
 Version 2.18.5
 --------------
 * When loading a simulation from a binary file, REBOUND now checks if the version of the binary file is the same as the current version.